MYR-189 refactor(fleet): address Claude Review — file split, expiry + test fixes
- CLAUDE.md 300-line rule (blocker): extract the two vehicle GETs into
  fleet_api_vehicle_get.go (fleet_api.go 301→229) and the status handler +
  response type into fleet_config_status_handler.go (fleet_config_handler.go
  328→277).
- daysRemaining is now nil once expired (UI branches on `expired` instead of
  a negative countdown); truncation semantics documented.
- stubFleetServer now registers t.Cleanup, closing the leaked inline test
  servers; redundant caller Close cleanups removed to avoid double-close.
- gosec pragma lists G107 + G704 (G704 is what the pinned gosec reports).
- Added a GET-path ownership-enforcement test.
